Rename API response class names as kenall-js
https://github.com/KEN-ALL/kenall-js/releases/tag/v1.0.5
