+++ src/Response/V1/AddressResolverResponse.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace zonuexe\Kenall\Response;
+namespace zonuexe\Kenall\Response\V1;
 
 use ArrayAccess;
 use ArrayIterator;
@@ -11,15 +11,16 @@ use JsonSerializable;
 use OutOfRangeException;
 use Psr\Http\Message\ResponseInterface;
 use Traversable;
+use zonuexe\Kenall\Response\ApiResponseInterface;
 use function json_decode;
 
 /**
  * Postal Code
  *
- * @template-implements ArrayAccess<int, Area>
- * @template-implements IteratorAggregate<int, Area>
+ * @template-implements ArrayAccess<int, Address>
+ * @template-implements IteratorAggregate<int, Address>
  */
-class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorAggregate, JsonSerializable
+class AddressResolverResponse implements ApiResponseInterface, ArrayAccess, IteratorAggregate, JsonSerializable
 {
     /** @var string */
     private $version;
@@ -45,12 +46,12 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
      *     town_chome: bool,
      *     town_multi: bool,
      *     town_raw: string,
-     *     corporation: array{
+     *     corporation: ?array{
      *         name: string,
      *         name_kana: string,
      *         block_lot: string,
      *         post_office: string,
-     *         code_type: string
+     *         code_type: int
      *     }
      * }>
      */
@@ -77,12 +78,12 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
      *     town_chome: bool,
      *     town_multi: bool,
      *     town_raw: string,
-     *     corporation: array{
+     *     corporation: ?array{
      *         name: string,
      *         name_kana: string,
      *         block_lot: string,
      *         post_office: string,
-     *         code_type: string
+     *         code_type: int
      *     }
      * }> $data
      */
@@ -101,12 +102,12 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
     }
 
     /**
-     * @return Traversable<int, Area>
+     * @return Traversable<int, Address>
      */
     public function getIterator()
     {
         foreach ($this->data as $i => $area) {
-            yield $i => new Area($area);
+            yield $i => Address::fromArray($area);
         }
     }
 
@@ -121,11 +122,11 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
 
     /**
      * @param int $offset
-     * @return Area
+     * @return Address
      */
     public function offsetGet($offset)
     {
-        return new Area($this->data[$offset]);
+        return Address::fromArray($this->data[$offset]);
     }
 
     /**
@@ -175,12 +176,12 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
          *         town_chome: bool,
          *         town_multi: bool,
          *         town_raw: string,
-         *         corporation: array{
+         *         corporation: ?array{
          *             name: string,
          *             name_kana: string,
          *             block_lot: string,
          *             post_office: string,
-         *             code_type: string
+         *             code_type: int
          *         }
          *     }>
          * } $data
@@ -211,12 +212,12 @@ class PostalCodeResponse implements ApiResponseInterface, ArrayAccess, IteratorA
      *     town_chome: bool,
      *     town_multi: bool,
      *     town_raw: string,
-     *     corporation: array{
+     *     corporation: ?array{
      *         name: string,
      *         name_kana: string,
      *         block_lot: string,
      *         post_office: string,
-     *         code_type: string
+     *         code_type: int
      *     }
      * }>}
      */
